Solution to Check Not Equal To

+++ Check Not Equal To@@67882a731b4392d8fc1577db/code.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 int main() {
-   int a,b;
-   scanf("%d %d",&a,&b);
+   float a,b;
+   scanf("%f %f",&a,&b);
    if(a!=b){
     printf("True");
    }
-   
+
    else{
     printf("False")
    }
